linux version

+++ src/parser_ragel.cpp
@@ -1,5 +1,5 @@
 
-/* #line 1 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 1 "ragel/parser.rl" */
 /*
  *  This file is part of the command-line tool sosicon.
  *  Copyright (C) 2012  Espen Andersen
@@ -24,7 +24,7 @@ namespace sosicon {
 
     //! \cond 
     
-/* #line 2 "D:\\Prosjekter\\sosicon\\src\\parser_ragel.cpp" */
+/* #line 28 "parser_ragel.cpp" */
 static const char _parseSosiLine_actions[] = {
 	0, 1, 0, 1, 2, 1, 4, 1, 
 	5, 1, 6, 2, 0, 1, 2, 0, 
@@ -129,7 +129,7 @@ static const int parseSosiLine_error = 0;
 static const int parseSosiLine_en_main = 7;
 
 
-/* #line 27 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 27 "ragel/parser.rl" */
 
     //! \endcond
 
@@ -157,12 +157,12 @@ ragelParseSosiLine( std::string sosiLine )
     int tmpint;
 
     
-/* #line 105 "D:\\Prosjekter\\sosicon\\src\\parser_ragel.cpp" */
+/* #line 161 "parser_ragel.cpp" */
 	{
 	cs = parseSosiLine_start;
 	}
 
-/* #line 108 "D:\\Prosjekter\\sosicon\\src\\parser_ragel.cpp" */
+/* #line 166 "parser_ragel.cpp" */
 	{
 	int _klen;
 	unsigned int _trans;
@@ -193,7 +193,7 @@ _resume:
 			else if ( (*p) > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (_mid - _keys);
+				_trans += (unsigned int)(_mid - _keys);
 				goto _match;
 			}
 		}
@@ -216,7 +216,7 @@ _resume:
 			else if ( (*p) > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += ((_mid - _keys)>>1);
+				_trans += (unsigned int)((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
@@ -237,7 +237,7 @@ _match:
 		switch ( *_acts++ )
 		{
 	case 0:
-/* #line 55 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 55 "ragel/parser.rl" */
 	{
             if( '\r' != (*p) ) {
                 tmpstr += (*p);
@@ -245,64 +245,64 @@ _match:
         }
 	break;
 	case 1:
-/* #line 61 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 61 "ragel/parser.rl" */
 	{
             tmpstr.clear();
         }
 	break;
 	case 2:
-/* #line 65 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 65 "ragel/parser.rl" */
 	{
             tmpint++;
         }
 	break;
 	case 3:
-/* #line 69 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 69 "ragel/parser.rl" */
 	{
             tmpint = 0;
         }
 	break;
 	case 4:
-/* #line 73 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 73 "ragel/parser.rl" */
 	{
             mPendingElementName = sosicon::utils::trim( tmpstr );
         }
 	break;
 	case 5:
-/* #line 77 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 77 "ragel/parser.rl" */
 	{
             mPendingElementAttributes = sosicon::utils::trim( tmpstr );
             tmpstr.clear();
         }
 	break;
 	case 6:
-/* #line 82 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 82 "ragel/parser.rl" */
 	{
             mPendingElementAttributes += ( " " + sosicon::utils::trim( tmpstr ) );
             tmpstr.clear();
         }
 	break;
 	case 7:
-/* #line 87 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 87 "ragel/parser.rl" */
 	{
             mPendingElementLevel = tmpint;
             tmpstr.clear();
         }
 	break;
 	case 8:
-/* #line 92 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 92 "ragel/parser.rl" */
 	{
             mPendingElementSerial = tmpstr.substr( 0, tmpstr.length() - 1 );
             tmpstr.clear();
         }
 	break;
 	case 9:
-/* #line 97 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 97 "ragel/parser.rl" */
 	{
             digestPendingElement();
         }
 	break;
-/* #line 237 "D:\\Prosjekter\\sosicon\\src\\parser_ragel.cpp" */
+/* #line 306 "parser_ragel.cpp" */
 		}
 	}
 
@@ -319,25 +319,25 @@ _again:
 	while ( __nacts-- > 0 ) {
 		switch ( *__acts++ ) {
 	case 1:
-/* #line 61 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 61 "ragel/parser.rl" */
 	{
             tmpstr.clear();
         }
 	break;
 	case 4:
-/* #line 73 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 73 "ragel/parser.rl" */
 	{
             mPendingElementName = sosicon::utils::trim( tmpstr );
         }
 	break;
 	case 6:
-/* #line 82 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 82 "ragel/parser.rl" */
 	{
             mPendingElementAttributes += ( " " + sosicon::utils::trim( tmpstr ) );
             tmpstr.clear();
         }
 	break;
-/* #line 268 "D:\\Prosjekter\\sosicon\\src\\parser_ragel.cpp" */
+/* #line 341 "parser_ragel.cpp" */
 		}
 	}
 	}
@@ -345,7 +345,7 @@ _again:
 	_out: {}
 	}
 
-/* #line 122 "D:\\Prosjekter\\sosicon\\src\\ragel\\parser.rl" */
+/* #line 122 "ragel/parser.rl" */
 
 
 };
